third_party/blink/web_tests/external/wpt/editing/run/insertparagraph_2001-3000-expected.txt

This is a testharness.js-based test.
Found 29 FAIL, 0 TIMEOUT, 0 NOTRUN.
[FAIL] [["defaultparagraphseparator","div"],["insertparagraph",""]] "<blockquote><p>foo[]<p>bar</blockquote>" compare innerHTML
  assert_equals: Unexpected innerHTML (after normalizing inline style) expected "<blockquote><p>foo</p><p><br></p><p>bar</p></blockquote>" but got "<blockquote><p>foo</p><p>bar</p></blockquote>"
[FAIL] [["defaultparagraphseparator","div"],["insertparagraph",""]] "<blockquote><p>foo[]<p>bar</blockquote>" queryCommandValue("defaultparagraphseparator") before
  assert_equals: Wrong result returned expected "p" but got "div"
[FAIL] [["defaultparagraphseparator","div"],["insertparagraph",""]] "<span>foo[]bar</span>" compare innerHTML
  assert_equals: Unexpected innerHTML (after normalizing inline style) expected "<div><span>foo</span></div><div><span>bar</span></div>" but got "<span>foo</span><div><span>bar</span></div>"
[FAIL] [["defaultparagraphseparator","p"],["insertparagraph",""]] "<span>foo[]bar</span>" compare innerHTML
  assert_equals: Unexpected innerHTML (after normalizing inline style) expected "<p><span>foo</span></p><p><span>bar</span></p>" but got "<span>foo</span><p><span>bar</span></p>"
[FAIL] [["defaultparagraphseparator","div"],["insertparagraph",""]] "<span>foo[]bar</span>baz" compare innerHTML
  assert_equals: Unexpected innerHTML (after normalizing inline style) expected "<div><span>foo</span></div><div><span>bar</span>baz</div>" but got "<span>foo</span><div><span>bar</span>baz</div>"
[FAIL] [["defaultparagraphseparator","p"],["insertparagraph",""]] "<span>foo[]bar</span>baz" compare innerHTML
  assert_equals: Unexpected innerHTML (after normalizing inline style) expected "<p><span>foo</span></p><p><span>bar</span>baz</p>" but got "<span>foo</span><p><span>bar</span>baz</p>"
[FAIL] [["stylewithcss","true"],["defaultparagraphseparator","div"],["insertparagraph",""]] "<b>foo[]bar</b>" compare innerHTML
  assert_equals: Unexpected innerHTML (after normalizing inline style) expected "<div><b>foo</b></div><div><b>bar</b></div>" but got "<b>foo</b><div><b>bar</b></div>"
[FAIL] [["stylewithcss","true"],["defaultparagraphseparator","div"],["insertparagraph",""]] "<b>foo[]bar</b>" queryCommandState("stylewithcss") before
  assert_equals: Wrong result returned expected false but got true
[FAIL] [["stylewithcss","false"],["defaultparagraphseparator","div"],["insertparagraph",""]] "<b>foo[]bar</b>" compare innerHTML
  assert_equals: Unexpected innerHTML (after normalizing inline style) expected "<div><b>foo</b></div><div><b>bar</b></div>" but got "<b>foo</b><div><b>bar</b></div>"
[FAIL] [["stylewithcss","true"],["defaultparagraphseparator","p"],["insertparagraph",""]] "<b>foo[]bar</b>" compare innerHTML
  assert_equals: Unexpected innerHTML (after normalizing inline style) expected "<p><b>foo</b></p><p><b>bar</b></p>" but got "<b>foo</b><p><b>bar</b></p>"
[FAIL] [["stylewithcss","false"],["defaultparagraphseparator","p"],["insertparagraph",""]] "<b>foo[]bar</b>" compare innerHTML
  assert_equals: Unexpected innerHTML (after normalizing inline style) expected "<p><b>foo</b></p><p><b>bar</b></p>" but got "<b>foo</b><p><b>bar</b></p>"
[FAIL] [["stylewithcss","true"],["defaultparagraphseparator","div"],["insertparagraph",""]] "<b>foo[]bar</b>baz" compare innerHTML
  assert_equals: Unexpected innerHTML (after normalizing inline style) expected "<div><b>foo</b></div><div><b>bar</b>baz</div>" but got "<b>foo</b><div><b>bar</b>baz</div>"
[FAIL] [["stylewithcss","false"],["defaultparagraphseparator","div"],["insertparagraph",""]] "<b>foo[]bar</b>baz" compare innerHTML
  assert_equals: Unexpected innerHTML (after normalizing inline style) expected "<div><b>foo</b></div><div><b>bar</b>baz</div>" but got "<b>foo</b><div><b>bar</b>baz</div>"
[FAIL] [["stylewithcss","true"],["defaultparagraphseparator","p"],["insertparagraph",""]] "<b>foo[]bar</b>baz" compare innerHTML
  assert_equals: Unexpected innerHTML (after normalizing inline style) expected "<p><b>foo</b></p><p><b>bar</b>baz</p>" but got "<b>foo</b><p><b>bar</b>baz</p>"
[FAIL] [["stylewithcss","false"],["defaultparagraphseparator","p"],["insertparagraph",""]] "<b>foo[]bar</b>baz" compare innerHTML
  assert_equals: Unexpected innerHTML (after normalizing inline style) expected "<p><b>foo</b></p><p><b>bar</b>baz</p>" but got "<b>foo</b><p><b>bar</b>baz</p>"
[FAIL] [["defaultparagraphseparator","div"],["insertparagraph",""]] "<b>foo[]</b>bar" compare innerHTML
  assert_equals: Unexpected innerHTML (after normalizing inline style) expected "<div><b>foo</b></div><div>bar</div>" but got "<b>foo</b><div>bar</div>"
[FAIL] [["defaultparagraphseparator","p"],["insertparagraph",""]] "<b>foo[]</b>bar" compare innerHTML
  assert_equals: Unexpected innerHTML (after normalizing inline style) expected "<p><b>foo</b></p><p>bar</p>" but got "<b>foo</b><p>bar</p>"
[FAIL] [["defaultparagraphseparator","div"],["insertparagraph",""]] "foo<b>[]bar</b>" compare innerHTML
  assert_equals: Unexpected innerHTML (after normalizing inline style) expected "<div>foo</div><div><b>bar</b></div>" but got "foo<div><b>bar</b></div>"
[FAIL] [["defaultparagraphseparator","p"],["insertparagraph",""]] "foo<b>[]bar</b>" compare innerHTML
  assert_equals: Unexpected innerHTML (after normalizing inline style) expected "<p>foo</p><p><b>bar</b></p>" but got "foo<p><b>bar</b></p>"
[FAIL] [["defaultparagraphseparator","div"],["insertparagraph",""]] "<b>foo[]</b><i>bar</i>" compare innerHTML
  assert_equals: Unexpected innerHTML (after normalizing inline style) expected "<div><b>foo</b></div><div><i>bar</i></div>" but got "<b>foo</b><div><i>bar</i></div>"
[FAIL] [["defaultparagraphseparator","p"],["insertparagraph",""]] "<b>foo[]</b><i>bar</i>" compare innerHTML
  assert_equals: Unexpected innerHTML (after normalizing inline style) expected "<p><b>foo</b></p><p><i>bar</i></p>" but got "<b>foo</b><p><i>bar</i></p>"
[FAIL] [["stylewithcss","true"],["defaultparagraphseparator","div"],["insertparagraph",""]] "<b id=x class=y>foo[]bar</b>" compare innerHTML
  assert_equals: Unexpected innerHTML (after normalizing inline style) expected "<div><b id=\\"x\\" class=\\"y\\">foo</b></div><div><b class=\\"y\\">bar</b></div>" but got "<b id=\\"x\\" class=\\"y\\">foo</b><div><b class=\\"y\\">bar</b></div>"
[FAIL] [["stylewithcss","false"],["defaultparagraphseparator","div"],["insertparagraph",""]] "<b id=x class=y>foo[]bar</b>" compare innerHTML
  assert_equals: Unexpected innerHTML (after normalizing inline style) expected "<div><b id=\\"x\\" class=\\"y\\">foo</b></div><div><b class=\\"y\\">bar</b></div>" but got "<b id=\\"x\\" class=\\"y\\">foo</b><div><b class=\\"y\\">bar</b></div>"
[FAIL] [["stylewithcss","true"],["defaultparagraphseparator","p"],["insertparagraph",""]] "<b id=x class=y>foo[]bar</b>" compare innerHTML
  assert_equals: Unexpected innerHTML (after normalizing inline style) expected "<p><b id=\\"x\\" class=\\"y\\">foo</b></p><p><b class=\\"y\\">bar</b></p>" but got "<b id=\\"x\\" class=\\"y\\">foo</b><p><b class=\\"y\\">bar</b></p>"
[FAIL] [["stylewithcss","false"],["defaultparagraphseparator","p"],["insertparagraph",""]] "<b id=x class=y>foo[]bar</b>" compare innerHTML
  assert_equals: Unexpected innerHTML (after normalizing inline style) expected "<p><b id=\\"x\\" class=\\"y\\">foo</b></p><p><b class=\\"y\\">bar</b></p>" but got "<b id=\\"x\\" class=\\"y\\">foo</b><p><b class=\\"y\\">bar</b></p>"
[FAIL] [["stylewithcss","true"],["defaultparagraphseparator","div"],["insertparagraph",""]] "<i><b>foo[]bar</b>baz</i>" compare innerHTML
  assert_equals: Unexpected innerHTML (after normalizing inline style) expected "<div><i><b>foo</b></i></div><div><i><b>bar</b>baz</i></div>" but got "<i><b>foo</b></i><div><i><b>bar</b>baz</i></div>"
[FAIL] [["stylewithcss","false"],["defaultparagraphseparator","div"],["insertparagraph",""]] "<i><b>foo[]bar</b>baz</i>" compare innerHTML
  assert_equals: Unexpected innerHTML (after normalizing inline style) expected "<div><i><b>foo</b></i></div><div><i><b>bar</b>baz</i></div>" but got "<i><b>foo</b></i><div><i><b>bar</b>baz</i></div>"
[FAIL] [["stylewithcss","true"],["defaultparagraphseparator","p"],["insertparagraph",""]] "<i><b>foo[]bar</b>baz</i>" compare innerHTML
  assert_equals: Unexpected innerHTML (after normalizing inline style) expected "<p><i><b>foo</b></i></p><p><i><b>bar</b>baz</i></p>" but got "<i><b>foo</b></i><p><i><b>bar</b>baz</i></p>"
[FAIL] [["stylewithcss","false"],["defaultparagraphseparator","p"],["insertparagraph",""]] "<i><b>foo[]bar</b>baz</i>" compare innerHTML
  assert_equals: Unexpected innerHTML (after normalizing inline style) expected "<p><i><b>foo</b></i></p><p><i><b>bar</b>baz</i></p>" but got "<i><b>foo</b></i><p><i><b>bar</b>baz</i></p>"
Harness: the test ran to completion.
